requirements.txt
asgiref==3.5.2
Django==4.1
django-cors-headers==3.13.0
django-filter==22.1
django-phonenumber-field==7.0.0
djangorestframework==3.13.1
phonenumbers==8.12.54
Pillow==9.2.0
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
